refactor: share MutableTexture test helper and simplify layout helpers
